[20/20] SPECfile cleanup - Do not try to re-define the RPM %license Virtual File Attribute to a global macro
https://rpm-software-management.github.io/rpm/manual/spec.html

This code was likely meant to deal %license tag on systems without RPM 4.11,
which could happen when the maintainer wanted to keep close the content
of Fedora SPECfile and RHEL SPECfile at a time.

While in Fedora the RPM 4.11 was present for years before the commit that added this code,
it might not be the case in an old, but still active, RHEL at that time.